SGHSS-VidaPlus/app/requirements.txt
Flask
Werkzeug
PyJWT
